Properly populate subscription ID column when sending MMS.
Populate the DB column, and feed the correct value into it for the !MSIM
case. Also clean up the code a little to make the subscription ID value
flow a little more readable.

+++ src/com/android/mms/transaction/PushReceiver.java
@@ -212,7 +212,7 @@ public class PushReceiver extends BroadcastReceiver {
 
                         if (!isDuplicateNotification(mContext, nInd)) {
                             int phoneId = intent.getIntExtra(PhoneConstants.SLOT_KEY, 0);
-                            long subId = intent.getLongExtra(PhoneConstants.SUBSCRIPTION_KEY, 0);
+                            int subId = intent.getIntExtra(PhoneConstants.SUBSCRIPTION_KEY, 0);
                             //Phone ID will be updated in data base
                             Log.d(TAG, "phoneId : " + phoneId + " subId : " + subId);
                             ContentValues values = new ContentValues(1);
